+++ src/main/scala/edu/ie3/simona/agent/grid/DBFSAlgorithm.scala
@@ -17,6 +17,7 @@ import edu.ie3.powerflow.model.NodeData.StateData
 import edu.ie3.powerflow.model.PowerFlowResult
 import edu.ie3.powerflow.model.PowerFlowResult.FailedPowerFlowResult.FailedNewtonRaphsonPFResult
 import edu.ie3.powerflow.model.PowerFlowResult.SuccessFullPowerFlowResult.ValidNewtonRaphsonPFResult
+import edu.ie3.powerflow.model.enums.NodeType
 import edu.ie3.simona.agent.grid.GridAgentData.{
   GridAgentBaseData,
   PowerFlowDoneData
@@ -38,6 +39,7 @@ import edu.ie3.simona.ontology.messages.SchedulerMessage.{
   ScheduleTriggerMessage,
   TriggerWithIdMessage
 }
+import edu.ie3.simona.ontology.messages.VoltageMessage.ProvideSlackVoltageMessage.ExchangeVoltage
 import edu.ie3.simona.ontology.messages.VoltageMessage.{
   ProvideSlackVoltageMessage,
   RequestSlackVoltageMessage
@@ -94,7 +96,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
       // 2. inferior grids p/q values
       askInferiorGridsForPowers(
         gridAgentBaseData.currentSweepNo,
-        gridAgentBaseData.gridEnv.subnetGateToActorRef,
+        gridAgentBaseData.gridEnv.subgridGateToActorRef,
         gridAgentBaseData.inferiorGridGates,
         gridAgentBaseData.powerFlowParams.sweepTimeout
       )
@@ -102,7 +104,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
       // 3. superior grids slack voltage
       askSuperiorGridsForSlackVoltages(
         gridAgentBaseData.currentSweepNo,
-        gridAgentBaseData.gridEnv.subnetGateToActorRef,
+        gridAgentBaseData.gridEnv.subgridGateToActorRef,
         gridAgentBaseData.superiorGridGates,
         gridAgentBaseData.powerFlowParams.sweepTimeout
       )
@@ -119,7 +121,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
         case receivedPowers: ReceivedPowerValues =>
           /* Can be a message from an asset or a message from an inferior grid */
           gridAgentBaseData.updateWithReceivedPowerValues(receivedPowers)
-        case receivedSlacks: ReceivedSlackValues =>
+        case receivedSlacks: ReceivedSlackVoltageValues =>
           gridAgentBaseData.updateWithReceivedSlackVoltages(receivedSlacks)
         case unknownReceivedValues =>
           throw new DBFSAlgorithmException(
@@ -157,19 +159,19 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
 
     // if we receive a request for slack voltages from our inferior grids we want to answer it
     case Event(
-          RequestSlackVoltageMessage(currentSweepNo, nodeUuid),
+          RequestSlackVoltageMessage(currentSweepNo, nodeUuids),
           gridAgentBaseData: GridAgentBaseData
         ) =>
       log.debug(
-        s"Received Slack Voltages request from {} for node {} and sweepNo: {}",
+        s"Received Slack Voltages request from {} for nodes {} and sweepNo: {}",
         sender(),
-        nodeUuid,
+        nodeUuids,
         gridAgentBaseData.currentSweepNo
       )
 
-      // we either have voltages ready calculated (not the first sweep) or we don't have them here
-      // -> return calculated value or target voltage as physical value
-      val (slackE, slackF) =
+      nodeUuids.map { nodeUuid =>
+        // we either have voltages ready calculated (not the first sweep) or we don't have them here
+        // -> return calculated value or target voltage as physical value
         (gridAgentBaseData.sweepValueStores.get(currentSweepNo) match {
           case Some(result) =>
             Some(result, currentSweepNo)
@@ -178,8 +180,8 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
             // the next sweep, as it triggers calculations for the next sweep or b) at all other
             // (non last downstream grid agents) in sweep 0
             log.debug(
-              "Unable to find slack voltage for node '{}' in sweep '{}'. Try to get voltage of previous sweep.",
-              nodeUuid,
+              "Unable to find slack voltage for nodes '{}' in sweep '{}'. Try to get voltage of previous sweep.",
+              nodeUuids,
               currentSweepNo
             )
             gridAgentBaseData.sweepValueStores
@@ -233,22 +235,25 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
             vSlack,
             Quantities.getQuantity(0, refSystemVUnit)
           )
+        } match {
+          case (slackE, slackF) =>
+            log.debug(
+              s"Provide {} to {} for node {} and sweepNo: {}",
+              s"$slackE, $slackF",
+              sender(),
+              nodeUuid,
+              gridAgentBaseData.currentSweepNo
+            )
+
+            ExchangeVoltage(nodeUuid, slackE, slackF)
         }
-
-      log.debug(
-        s"Provide {} to {} for node {} and sweepNo: {}",
-        s"$slackE, $slackF",
-        sender(),
-        nodeUuid,
-        gridAgentBaseData.currentSweepNo
-      )
-
-      stay() replying ProvideSlackVoltageMessage(
-        currentSweepNo,
-        nodeUuid,
-        slackE,
-        slackF
-      )
+      } match {
+        case exchangeVoltages =>
+          stay() replying ProvideSlackVoltageMessage(
+            currentSweepNo,
+            exchangeVoltages
+          )
+      }
 
     // receive grid power values message request from superior grids
     // / before power flow calc for this sweep we either have to stash() the message to answer it later (in current sweep)
@@ -278,61 +283,104 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
 
     // / after power flow calc for this sweepNo
     case Event(
-          RequestGridPowerMessage(_, requestedNodeUuid),
-          PowerFlowDoneData(gridAgentBaseData, powerFlowResult)
+          RequestGridPowerMessage(_, requestedNodeUuids),
+          powerFlowDoneData @ PowerFlowDoneData(
+            gridAgentBaseData,
+            powerFlowResult,
+            pendingRequestAnswers
+          )
         ) =>
-      log.debug(
-        "Received request for grid power values and im READY to provide."
+      /* Determine the subgrid number of the grid agent, that has sent the request */
+      val firstRequestedNodeUuid = requestedNodeUuids.headOption.getOrElse(
+        throw new DBFSAlgorithmException(
+          "Did receive a grid power request but without specified nodes"
+        )
       )
+      gridAgentBaseData.gridEnv.subgridGateToActorRef
+        .map { case (subGridGate, _) => subGridGate.getSuperiorNode }
+        .find(_.getUuid == firstRequestedNodeUuid)
+        .map(_.getSubnet) match {
+        case Some(requestingSubgridNumber) =>
+          powerFlowResult match {
+            case validNewtonRaphsonPFResult: ValidNewtonRaphsonPFResult =>
+              val exchangePowers = requestedNodeUuids
+                .map { nodeUuid =>
+                  /* Figure out the node index for each requested node */
+                  nodeUuid -> gridAgentBaseData.gridEnv.gridModel.nodeUuidToIndexMap
+                    .get(nodeUuid)
+                    .flatMap { nodeIndex =>
+                      /* Find matching node result */
+                      validNewtonRaphsonPFResult.nodeData.find(stateData =>
+                        stateData.index == nodeIndex
+                      )
+                    }
+                    .map {
+                      case StateData(_, nodeType, _, power)
+                          if nodeType == NodeType.SL =>
+                        val refSystem =
+                          gridAgentBaseData.gridEnv.gridModel.mainRefSystem
+                        val (pInPu, qInPu) =
+                          (power.real, power.imag)
+                        // The power flow result data provides the nodal residual power at the slack node.
+                        // A feed-in case from the inferior grid TO the superior grid leads to positive residual power at the
+                        // inferior grid's *slack node* (superior grid seems to be a load to the inferior grid).
+                        // To model the exchanged power from the superior grid's point of view, -1 has to be multiplied.
+                        // (Inferior grid is a feed in facility to superior grid, which is negative then). Analogously for load case.
+                        (
+                          refSystem.pInSi(pInPu).multiply(-1),
+                          refSystem.qInSi(qInPu).multiply(-1)
+                        )
+                      case _ =>
+                        /* TODO: As long as there are no multiple slack nodes, provide "real" power only for the slack node */
+                        (
+                          Quantities.getQuantity(0d, MEGAWATT),
+                          Quantities.getQuantity(0d, MEGAVAR)
+                        )
+                    }
+                    .getOrElse {
+                      throw new DBFSAlgorithmException(
+                        s"Got a request for power @ node with uuid $requestedNodeUuids but cannot find it in my result data!"
+                      )
+                    }
+                }
+                .map { case (nodeUuid, (p, q)) =>
+                  ProvideGridPowerMessage.ExchangePower(nodeUuid, p, q)
+                }
 
-      // get the index from the power flow data of the requested node
-      val requestedNodeIdxOpt =
-        gridAgentBaseData.gridEnv.gridModel.nodeUuidToIndexMap
-          .get(requestedNodeUuid)
+              /* Determine the remaining replies */
+              val stillPendingRequestAnswers =
+                pendingRequestAnswers.filterNot(_ == requestingSubgridNumber)
 
-      powerFlowResult match {
-        case validNewtonRaphsonPFResult: ValidNewtonRaphsonPFResult =>
-          val (p, q) = validNewtonRaphsonPFResult.nodeData.find(stateData =>
-            requestedNodeIdxOpt.contains(stateData.index)
-          ) match {
-            case Some(apparentPowerS) =>
-              val refSystem = gridAgentBaseData.gridEnv.gridModel.mainRefSystem
-              val (pInPu, qInPu) =
-                (apparentPowerS.power.real, apparentPowerS.power.imag)
-              // The power flow result data provides the nodal residual power at the slack node.
-              // A feed-in case from the inferior grid TO the superior grid leads to positive residual power at the
-              // inferior grid's *slack node* (superior grid seems to be a load to the inferior grid).
-              // To model the exchanged power from the superior grid's point of view, -1 has to be multiplied.
-              // (Inferior grid is a feed in facility to superior grid, which is negative then). Analogously for load case.
-              (
-                refSystem.pInSi(pInPu).multiply(-1),
-                refSystem.qInSi(qInPu).multiply(-1)
-              )
-            case None =>
-              throw new DBFSAlgorithmException(
-                s"Got a request for power @ node with uuid $requestedNodeUuid but cannot find it in my result data!"
-              )
+              // update the sweep value store and clear all received maps
+              // note: normally it is expected that this has to be done after power flow calculations but for the sake
+              // of having it only once in the code we put this here. Otherwise it would have to been put before EVERY
+              // return with a valid power flow result (currently happens already in two situations)
+              val updatedGridAgentBaseData =
+                if (stillPendingRequestAnswers.isEmpty) {
+                  gridAgentBaseData.storeSweepDataAndClearReceiveMaps(
+                    validNewtonRaphsonPFResult,
+                    gridAgentBaseData.superiorGridNodeUuids,
+                    gridAgentBaseData.inferiorGridGates
+                  )
+                } else {
+                  powerFlowDoneData.copy(pendingRequestAnswers =
+                    stillPendingRequestAnswers
+                  )
+                }
+
+              stay() replying
+                ProvideGridPowerMessage(
+                  exchangePowers
+                ) using updatedGridAgentBaseData
+
+            case _: FailedNewtonRaphsonPFResult =>
+              stay() replying FailedPowerFlow using gridAgentBaseData
           }
-
-          // update the sweep value store and clear all received maps
-          // note: normally it is expected that this has to be done after power flow calculations but for the sake
-          // of having it only once in the code we put this here. Otherwise it would have to been putted before EVERY
-          // return with a valid power flow result (currently happens already in two situations)
-          val updatedGridAgentBaseData =
-            gridAgentBaseData.storeSweepDataAndClearReceiveMaps(
-              validNewtonRaphsonPFResult,
-              gridAgentBaseData.superiorGridNodeUuids,
-              gridAgentBaseData.inferiorGridGates
-            )
-
-          stay() replying
-            ProvideGridPowerMessage(
-              requestedNodeUuid,
-              p,
-              q
-            ) using updatedGridAgentBaseData
-
-        case _: FailedNewtonRaphsonPFResult =>
+        case None =>
+          /* It is not possible to determine, who has asked */
+          log.error(
+            "I got a grid power request from a subgrid I don't know. Can't answer it properly."
+          )
           stay() replying FailedPowerFlow using gridAgentBaseData
       }
 
@@ -347,7 +395,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
       // request the updated slack voltages from the superior grid
       askSuperiorGridsForSlackVoltages(
         gridAgentBaseData.currentSweepNo,
-        gridAgentBaseData.gridEnv.subnetGateToActorRef,
+        gridAgentBaseData.gridEnv.subgridGateToActorRef,
         gridAgentBaseData.superiorGridGates,
         gridAgentBaseData.powerFlowParams.sweepTimeout
       )
@@ -365,8 +413,9 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
       // inform my child grids about the end of this grid simulation
       gridAgentBaseData.inferiorGridGates
         .map {
-          gridAgentBaseData.gridEnv.subnetGateToActorRef(_)
+          gridAgentBaseData.gridEnv.subgridGateToActorRef(_)
         }
+        .distinct
         .foreach(_ ! FinishGridSimulationTrigger(currentTick))
 
       // inform every system participant about the end of this grid simulation
@@ -566,7 +615,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
     // this means we requested an update of the slack voltage values, but for now don't request (and hence don't expect)
     // updated power values for our power flow calculations
     case Event(
-          receivedSlackValues: ReceivedSlackValues,
+          receivedSlackValues: ReceivedSlackVoltageValues,
           gridAgentBaseData: GridAgentBaseData
         ) =>
       log.debug(
@@ -632,7 +681,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
           val askForInferiorGridPowersOpt =
             askInferiorGridsForPowers(
               gridAgentBaseData.currentSweepNo,
-              gridAgentBaseData.gridEnv.subnetGateToActorRef,
+              gridAgentBaseData.gridEnv.subgridGateToActorRef,
               gridAgentBaseData.inferiorGridGates,
               gridAgentBaseData.powerFlowParams.sweepTimeout
             )
@@ -1095,7 +1144,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
   private def askInferiorGridsForPowers(
       currentSweepNo: Int,
       subGridGateToActorRef: Map[SubGridGate, ActorRef],
-      inferiorGridGates: Vector[SubGridGate],
+      inferiorGridGates: Seq[SubGridGate],
       askTimeout: Duration
   ): Option[Future[ReceivedPowerValues]] = {
     implicit val timeout: AkkaTimeout = AkkaTimeout.create(askTimeout)
@@ -1112,10 +1161,17 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
                 inferiorGridGate
               ) -> inferiorGridGate.getSuperiorNode.getUuid
             }
-            .map { case (inferiorGridAgentRef, inferiorGridGateNode) =>
+            .groupMap {
+              // Group the gates by target actor, so that only one request is sent per grid agent
+              case (inferiorGridAgentRef, _) =>
+                inferiorGridAgentRef
+            } { case (_, inferiorGridGates) =>
+              inferiorGridGates
+            }
+            .map { case (inferiorGridAgentRef, inferiorGridGateNodes) =>
               (inferiorGridAgentRef ? RequestGridPowerMessage(
                 currentSweepNo,
-                inferiorGridGateNode
+                inferiorGridGateNodes.distinct
               )).map {
                 case provideGridPowerMessage: ProvideGridPowerMessage =>
                   (inferiorGridAgentRef, provideGridPowerMessage)
@@ -1123,6 +1179,7 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
                   (inferiorGridAgentRef, FailedPowerFlow)
               }
             }
+            .toVector
         )
         .map(ReceivedGridPowerValues)
         .pipeTo(self)
@@ -1148,26 +1205,29 @@ trait DBFSAlgorithm extends PowerFlowSupport with GridResultsSupport {
       subGridGateToActorRef: Map[SubGridGate, ActorRef],
       superiorGridGates: Vector[SubGridGate],
       askTimeout: Duration
-  ): Option[Future[ReceivedSlackValues]] = {
+  ): Option[Future[ReceivedSlackVoltageValues]] = {
     implicit val timeout: AkkaTimeout = AkkaTimeout.create(askTimeout)
     log.debug(
       s"asking superior grids for slack voltage values: {}",
       superiorGridGates
     )
+
     Option.when(superiorGridGates.nonEmpty) {
       Future
         .sequence(
-          superiorGridGates.map(superiorGridGate => {
-            val superiorGridAgent = subGridGateToActorRef(superiorGridGate)
-            (superiorGridAgent ? RequestSlackVoltageMessage(
-              currentSweepNo,
-              superiorGridGate.getSuperiorNode.getUuid
-            )).map { case providedSlackValues: ProvideSlackVoltageMessage =>
-              (superiorGridAgent, providedSlackValues)
+          superiorGridGates
+            .groupBy(subGridGateToActorRef(_))
+            .map { case (superiorGridAgent, gridGates) =>
+              (superiorGridAgent ? RequestSlackVoltageMessage(
+                currentSweepNo,
+                gridGates.map(_.getSuperiorNode.getUuid)
+              )).map { case providedSlackValues: ProvideSlackVoltageMessage =>
+                (superiorGridAgent, providedSlackValues)
+              }
             }
-          })
+            .toVector
         )
-        .map(ReceivedSlackValues)
+        .map(ReceivedSlackVoltageValues)
         .pipeTo(self)
     }
   }