+++ src/test/scala/edu/ie3/simona/model/grid/GridSpec.scala
@@ -100,7 +100,7 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
 
     "validate the connectivity of a connected grid correctly" in new BasicGridWithSwitches {
       // enable nodes
-      override val nodes: Set[NodeModel] = super.nodes
+      override val nodes: Seq[NodeModel] = super.nodes
       nodes.foreach(_.enable())
 
       // enable lines
@@ -135,7 +135,7 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
 
     "throw an InvalidGridException if a grid is not connected" in new BasicGridWithSwitches {
       // enable nodes
-      override val nodes: Set[NodeModel] = super.nodes
+      override val nodes: Seq[NodeModel] = super.nodes
       nodes.foreach(_.enable())
 
       // remove a line from the grid
@@ -170,7 +170,7 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
 
     "throw an InvalidGridException if two switches are connected @ the same node" in new BasicGridWithSwitches {
       // enable nodes
-      override val nodes: Set[NodeModel] = super.nodes
+      override val nodes: Seq[NodeModel] = super.nodes
       nodes.foreach(_.enable())
 
       // add a second switch @ node13 (between node1 and node13)
@@ -219,7 +219,7 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
 
       "contains 3 open switches" in new BasicGridWithSwitches {
         // enable nodes
-        override val nodes: Set[NodeModel] = super.nodes
+        override val nodes: Seq[NodeModel] = super.nodes
         nodes.foreach(_.enable())
         // enable switches
         override val switches: Set[SwitchModel] = super.switches
@@ -264,7 +264,7 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
 
       "contains 3 closed switches" in new BasicGridWithSwitches {
         // enable nodes
-        override val nodes: Set[NodeModel] = super.nodes
+        override val nodes: Seq[NodeModel] = super.nodes
         nodes.foreach(_.enable())
         // enable switches
         override val switches: Set[SwitchModel] = super.switches
@@ -349,7 +349,7 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
 
       "contains 1 open and 2 closed switches" in new BasicGridWithSwitches {
         // enable nodes
-        override val nodes: Set[NodeModel] = super.nodes
+        override val nodes: Seq[NodeModel] = super.nodes
         nodes.foreach(_.enable())
         // enable switches
         override val switches: Set[SwitchModel] = super.switches
@@ -423,7 +423,7 @@ class GridSpec extends UnitSpec with LineInputTestData with DefaultTestData {
 
       "contains no switches" in new BasicGrid {
         // enable nodes
-        override val nodes: Set[NodeModel] = super.nodes
+        override val nodes: Seq[NodeModel] = super.nodes
         nodes.foreach(_.enable())
 
         // get the grid from the raw data